+++ src/core/messaging/ChatAdapterConfig.ts
@@ -3,4 +3,6 @@ export default interface ChatAdapterConfig {
     webChatIC3AdapterCDNUrl?: string;
     webChatACSAdapterVersion?: string;
     webChatACSAdapterCDNUrl?: string;
+    webChatDirectLineVersion?: string;
+    webChatDirectLineCDNUrl?: string;
 }