+++ src/core/messaging/ChatAdapterOptionalParams.ts
@@ -4,12 +4,17 @@ interface ChatAdapterOptionalParams {
         options?: {
             [key: string]: any // eslint-disable-line @typescript-eslint/no-explicit-any
         }
-    }
+    };
     ACSAdapter?: {
         options?: {
             [key: string]: any // eslint-disable-line @typescript-eslint/no-explicit-any
         }
-    }
+    };
+    DirectLine?: {
+        options?: {
+            [key: string]: any // eslint-disable-line @typescript-eslint/no-explicit-any
+        }
+    };
 }
 
 export default ChatAdapterOptionalParams;